We met on social media platform and started working together, tech guy was making made a Management girl, then finally our first meeting was at isckon temple mumbai, and we started this beutiful relationship with Blessings of Radhe-Krishna, reason behid getting married in triyuginarayan temple is we both love shiva and tiryuginayarayan is the place where Shiv-parvati got married.
Loving each other is a very comman thing in this world but no one isnt perfect in this whole world, we not just love each others for good and happy moments we also accepted each others flaws and share difficult times, we hold each others hands in both the situations. We started this journey woth radhe krishna's blessing and getting married at Shiv parvati's marrage place, we would love to see you in this event beutiful step of both of our jouney.

Reframe this and add this in our story section